Moved stylesheets to separate folder & styled create trip page to match other pages
