+++ client/src/App.js
@@ -2,10 +2,10 @@ import React, { Component } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
 import 'bootstrap/dist/css/bootstrap.min.css';
-import './App.css'
+import './Stylesheets/App.css'
 
-import CreateAccount from "./components/account/CreateAccount";
-import SignIn from "./components/account/SignIn";
+import CreateAccount from "./components/Account/CreateAccount";
+import SignIn from "./components/Account/SignIn";
 import Home from "./routes/Home.js"
 
 class App extends Component {
