+++ client/src/Stylesheets/CreateAccount.css
@@ -8,7 +8,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    background-image: url("../../Media//signupImage.svg");
+    background-image: url("../Media/signupImage.svg");
     background-repeat: no-repeat;
     background-size: contain;
     background-position: right top;
