+++ client/src/Stylesheets/Trip.css
