+++ client/src/Stylesheets/Trips.css
